changed amount unit from int to long

+++ protocore/src/main/java/com/satispay/protocore/models/payment/MealVoucher.java
@@ -4,16 +4,16 @@ public class MealVoucher {
 
     public MealVoucher() {}
 
-    public int max_amount_unit;
+    public long max_amount_unit;
 
     public int max_number;
 
-    public MealVoucher(int maxAmountUnit, int maxNumber) {
+    public MealVoucher(long maxAmountUnit, int maxNumber) {
         max_amount_unit = maxAmountUnit;
         max_number = maxNumber;
     }
 
-    public int getMaxAmountUnit() { return max_amount_unit; }
+    public long getMaxAmountUnit() { return max_amount_unit; }
 
     public int getMaxNumber() { return max_number; }
